File: move color file to theme package
