+++ core/designsystem/src/main/java/com/kolown/designsystem/component/PorringTextField.kt
@@ -37,13 +37,13 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
-import com.kolown.designsystem.Error
-import com.kolown.designsystem.Primary
-import com.kolown.designsystem.PrimaryDark
-import com.kolown.designsystem.PrimaryUnActive
-import com.kolown.designsystem.PrimaryUnActiveDark
-import com.kolown.designsystem.Surface
-import com.kolown.designsystem.SurfaceError
+import com.kolown.designsystem.ui.theme.Error
+import com.kolown.designsystem.ui.theme.Primary
+import com.kolown.designsystem.ui.theme.PrimaryDark
+import com.kolown.designsystem.ui.theme.PrimaryUnActive
+import com.kolown.designsystem.ui.theme.PrimaryUnActiveDark
+import com.kolown.designsystem.ui.theme.Surface
+import com.kolown.designsystem.ui.theme.SurfaceError
 
 @Composable
 fun PorringTextField(
