+++ feature/login/src/main/java/com/kolown/login/component/ButtonWithIcon.kt
@@ -14,8 +14,8 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.kolown.designsystem.Primary
-import com.kolown.designsystem.Surface
+import com.kolown.designsystem.ui.theme.Primary
+import com.kolown.designsystem.ui.theme.Surface
 
 @Composable
 internal fun ButtonWithIcon(
