+++ feature/main/src/main/java/com/kolown/main/component/MainBottomBar.kt
@@ -24,8 +24,8 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.kolown.designsystem.Primary
-import com.kolown.designsystem.PrimaryUnActive
+import com.kolown.designsystem.ui.theme.Primary
+import com.kolown.designsystem.ui.theme.PrimaryUnActive
 import com.kolown.main.navigation.MainMenu
 import com.kolown.main.ui.theme.PorringTheme
 import com.kolown.navigation.MainMenuRoute
