+++ feature/my/src/main/java/com/kolown/my/MyScreen.kt
@@ -38,8 +38,8 @@ import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import coil3.compose.AsyncImage
 import com.kolown.common.component.RestrictedLoginContent
-import com.kolown.designsystem.Primary
 import com.kolown.designsystem.R
+import com.kolown.designsystem.ui.theme.Primary
 import com.kolown.model.PostContentModel
 import com.kolown.my.component.GalleryItem
 import com.kolown.my.component.MyAppBar
