+++ feature/my/src/main/java/com/kolown/my/component/GalleryItem.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import com.kolown.data.mock.MockDataProvider
-import com.kolown.designsystem.Surface2
+import com.kolown.designsystem.ui.theme.Surface2
 import com.kolown.model.PostContentModel
 import kotlin.math.absoluteValue
 import kotlin.random.Random
