+++ feature/search/src/main/java/com/kolown/search/component/TagSearchBar.kt
@@ -1,32 +1,17 @@
 package com.kolown.search.component
 
-import android.util.Log
-import androidx.compose.animation.animateContentSize
-import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.SearchBar
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.onFocusChanged
@@ -37,10 +22,9 @@ import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
-import com.kolown.designsystem.Primary
-import com.kolown.designsystem.Surface2
+import com.kolown.designsystem.ui.theme.Primary
+import com.kolown.designsystem.ui.theme.Surface2
 import com.kolown.search.R
 
 @Composable
@@ -51,7 +35,7 @@ fun TagSearchBar(
     focusState: Boolean,
     onFocusChange: (Boolean) -> Unit = {},
     onClearClick: () -> Unit = {},
-    onBackButtonClicked: () -> Unit = {}
+    onBackButtonClicked: () -> Unit = {},
 ) {
 
     Column(
@@ -94,8 +78,8 @@ fun TagSearchBar(
                 }
             },
             trailingIcon = {
-                if(focusState){
-                    IconButton(onClick = onClearClick){
+                if (focusState) {
+                    IconButton(onClick = onClearClick) {
                         Icon(
                             imageVector = Icons.Default.Close,
                             contentDescription = "search_back", tint = Primary
